refactor RestaurantUserRestaurant POJO to entities
